:boom:[HARDWARE REWORK] Fix reset button boot fail
Reset settings button is supposed to be pressed on boot,
but since it is on GPIO0, ESP8266 enters in boot mode.

The issue could be solved by software changing the logic allowing reset
only after the boot, but this sounds dangerous under several aspects.

The adopted solution requires hardware changes and uses the DMX buffers
enable pin to sense reset button at boot.

The PCB rework consist of:
- cut the path conneting GPIO0 and the reset button
- disconnect the button's debouncing capacitor from GPIO0
- connect the button and its debouncing capacitor to GPIO4

Newer version of the board will fix the issue avoiding board reworking.